Try to get pywasmbuild entrypoint working
