+++ pyodide-build/pyproject.toml
@@ -4,4 +4,4 @@ requires = ["setuptools>=42", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [entry_points]
-console_scripts = ["pywasmbuild = __main__:out_of_tree_main"]
+console_scripts = ["pywasmbuild = out_of_tree_main:out_of_tree_main"]
