+++ pyodide-build/setup.py
@@ -3,8 +3,7 @@ import setuptools
 
 if __name__ == "__main__":
     setuptools.setup(
-        # entry_points = {
-        #     "console_scripts" :
-        #     ["foo = foomod:main"]
-        # }
+        entry_points = {
+            "console_scripts" : ["pywasmbuild = pyodide_build.out_of_tree_main:out_of_tree_main"]
+        }
     )
